fix(catalog): keep server-only automation hooks out of the client bundle
The P4 entity-changed hook added `await import('@/lib/automations/emit')` (which
imports `server-only`) to CatalogEntities. Because the scorecards `rule-builder`
(reachable from the client RuleBuilder via ScorecardForm) imported the
ENTITY_KINDS/RELATION_TYPES constants through the `@/collections/catalog` barrel,
dev webpack (no tree-shaking) pulled the whole collection config — and now its
server-only hook dep — into the browser bundle, 500-ing /dashboard, /scorecards,
and scorecard detail pages with: "You're importing a component that needs
'server-only'".

Fix: extract the vocabularies into a framework-light
`collections/catalog/constants.ts` (no Payload/hook imports) and point the
client-reachable `rule-builder` at it directly; the collection files and the
barrel re-export from there for back-compat.

Verified: tsc clean (baseline 110), 32 scorecards tests pass, lint clean; logged
in and loaded /dashboard, /scorecards, and /scorecards/<id> (the failing route) —
all 200, no build-error overlay, no server-only errors in the dev log.
